requirements.txt
flask-restplus==0.10.1
Flask-SQLAlchemy==2.2
musicbrainzngs==0.6

